Special-case char RET to allow for jumping to end-of-lines.
With this change you can use `avy-goto-char' also to jump to the end of
some currently visible line.  E.g., C-c j <return> (assuming C-c j is
bound to `avy-goto-char') makes all line endings jump targets.

`avy--overlay-at' had to be adapted so that the overlay at \n doesn't
make the line-break disappear.

`avy--overlay' had to be adapted in order not to put an overlay on the
char after (point-max) in case `avy--overlay-post' is used.  (Honestly,
this feature is a bit awkward with 'post overlays where the jump
location is visualized as first char in the next line.)