Add avy-action-ispell to avy.el
Thank you for creating avy-mode! It is really fun and fast to use. I thought it would be fitting for me to add a convenience function to make spell checking less of a chore. The point should return to its starting location after the desired word is corrected. As is the case for the other avy-action-{something} functions, this function is also added to the avy-dispatch-alist. I hope you'll consider my pull request worthy of  a merge. Thanks again!

Regards,

Emmanuel