Allow candidate face painting direction to be set by user.
The traditional 'right to left' painting makes visual seeking much
slower because the first character is one color for length 1
candidates, another color for length 2 candidates and still another
color for length 3 candidates. By painting left to right the first
character will always be the same color, and similarly so for 2nd
and 3rd.